+++ Game/StartNewLevelScreen.cpp
@@ -135,7 +135,7 @@ void StartNewLevelScreen::handleMouseWheelInput(SDL_MouseWheelEvent e) {
 
 	int heightOfScrolBlock = 0;
 
-	heightOfScrolBlock += _storyText->textLines.size() * 25;
+	heightOfScrolBlock += (int)_storyText->textLines.size() * 25;
 	heightOfScrolBlock += 120;
 
 	int currentY = _storyTitle->rect.y;
